Added second cache system that does not directly work on an asset but instead saves the canvas at a specified interval and at NodeEditor shutdown
This enables future undo support (Undo does not work on assets) and is significantly faster, especially when creating nodes.
But is not completely safe when Unity crashes, as it will revert to the last cache save, depending on the set cache interval (default 1min).
A small stutter occurs when caching, so it should not be too frequent, but caching is prohibited while user performs an action (dragging,...) so it's not always noticeable.

It is possible to revert back to the old cache system if desired by uncommenting the first line '#define EDITOR_CACHE_ASSET' in NodeEditorUserCache.cs