#70 Modular Canvas and Traversal behaviours (#109)
#70 Modular Canvas and Traversal behaviours

This update streamlines the process of adding new types of canvases with new properties, behaviours, requirements and traversal algorithms to the framework in a modular way.
Canvas behaviour can be scripted by controlling many aspects of the canvas and it's modifications.
It's also possible to add new ways of traversing these canvases, either similar to the default calculation, a statemachine, a dialogue system, or just a data visualizer, this is now possible without workarounds.

This is made possible by having a canvas type specify the basic properties and behaviours and a seperate traversal class manage the traversal of the canvas. Many canvas types could use the same traversal algorithm.

WARNING:
I seperated the default calculation behaviour from canvas base class, making it a seperate subclass, and NodeCanvas now serves only as a base class. 
Due to compability with previous save files (which would be broken if NodeCanvas was now abstract) it's now weakly limited (canvases with default type are possible but prohibited). 
In order to use previous saves, you first have to specify a valid type (any Canvas subclass) after loading. Also note later on NodeCanvas WILL be made abstract and thus older saves will be broken!

Included is a Graph Canvas type with seperate (yet empty) traversal behaviour to show how easily canvases can be scripted.

Changelog:
- NodeCanvas is now the base class for every other canvas type
- Added warning and side panel entry to convert canvas if the loaded canvas is of type NodeCanvas (base type). Rest of the framework is also mostly disabled if current canvas is of base type
- Added NodeCanvasTraversal base class to add unique traversal algorithms to a specific canvas type
- Added various callbacks to NodeCanvas like OnCreate, OnValidate, OnBeforeSavingCanvas, CanAddNode
- Added NodeCanvas.CreateCanvas methods to create a canvas
- NodeEditor is now completely lifted off the calculation code, API to calculate changed to `nodeCanvas.TraverseAll ()` and `nodeCanvas.OnNodeChange (Node)`
- Added Graph canvas example
- Fixed warnings in RootGraphNode