Implemented Undo functionality with UndoPro
Seems to be solid, as far as I can tell
Supports:
- Node GUI
- Node Selection / Dragging
- Connection Creation/Deletion (including affected connections)
- Node Creation / Duplication / Deletion
- Records survive Playmode/Compilation, not scene switch (as intended)
Does NOT support:
- Node Groups (does anyone actually use them?)

ATTENTION! Modifications done to support Undo:
Undo system is REFERENCE-based, no ScriptableObjects (SO) actually gets deleted until a new canvas is loaded / the scene is switched. So SOs (even if deleted) need to survive Compilation/Playmode. Currently, a working copy is saved and again loaded afterwards. That had to change. Now, it stores a 'SO memory dump' as the CurSession, which includes current and old SOs used for undo. Working-Copy save is only used for crash-critical saves. These include all but the 'window lost focus save', which is actually faster now.

NodeEditorUndoActions only serves to simplify the actual Undo calls and hides the error-checking.