Fixed temp folder location when installed through Unity Package Manager
Added new system to manage temp folder.
On normal install in asset folder, it will be similar to before, in the editorPath (usually Assets/Plugins/Node_Editor_Framework)
For Unity Package Manager Distribution, a new folder in the Asset database was needed.
Currently, it defaults to DefaultEditorResources/Node_Editor_Framework/, but it can be easily changed by moving the marker file along with the session files to a new folder.
Also this system gives full control to 3rd party tools basing off of Node Editor Framework, both embedded and through Unity Package Manager.
When creating a custom window for their tools, they can choose where to put their session files or to share them with the default window.
More information on that in NodeEditorWindow.cs