Added Framework extends example
The newest packaging in an assembly made it hard for 3rd party tools to extend the framework.
Especially if they do not have access to the source (e.g. when Node_Editor_Framework is a shared install, using Unity Package Manager)
This example show how the extension of existing fixed classes using the partial class feature of C# works even when the assemblies are separated.
See ExtendsUsage.md for more

Note: Of course normal extension using custom NodeCanvases, Nodes, etc. still work just fine. This is for when you want to add custom state data to NodeEditorState, for example.